Add files via upload
 the node struct represents a node in the binary search tree, with three fields: data, left, and right. The data field stores the value of the node, while the left and right fields point to the left and right subtrees, respectively.

The createNode function creates a new node with the specified data value and returns a pointer to the new node. The insertNode function inserts a new node with the specified data value into the tree rooted at root, following the rules of a binary search tree. The postorderTraversal function performs a postorder traversal of the tree rooted at root, printing out the data values of the nodes in postorder.

In the main function, we first create an empty binary search tree by setting the root pointer to NULL. We then insert some nodes into the tree using the insertNode function, and finally perform a postorder traversal of the tree using the postorderTraversal function.